[actuator] Move ApplyField and build_apply_lambda_action to actuator component
Centralises the folded-lambda action helper in actuator/__init__.py so both
cover.control and valve.control use the same implementation. Removes ~70 lines
of duplication from valve/__init__.py and eliminates orphaned imports from both
cover and valve.
