+++ esphome/components/cover/__init__.py
@@ -1,5 +1,3 @@
-from collections.abc import Callable
-from dataclasses import dataclass
 import logging
 
 from esphome import automation
@@ -39,14 +37,14 @@ from esphome.const import (
     DEVICE_CLASS_SHUTTER,
     DEVICE_CLASS_WINDOW,
 )
-from esphome.core import CORE, ID, CoroPriority, Lambda, coroutine_with_priority
+from esphome.core import CORE, ID, CoroPriority, coroutine_with_priority
 from esphome.core.entity_helpers import (
     entity_duplicate_validator,
     queue_entity_register,
     setup_device_class,
     setup_entity,
 )
-from esphome.cpp_generator import LambdaExpression, MockObj, MockObjClass
+from esphome.cpp_generator import MockObj, MockObjClass
 from esphome.types import ConfigType, TemplateArgsType
 
 IS_PLATFORM_COMPONENT = True
@@ -303,84 +301,13 @@ COVER_CONTROL_ACTION_SCHEMA = cv.Schema(
 )
 
 
-@dataclass(frozen=True)
-class ApplyField:
-    """One field in a folded-lambda action.
-
-    `conf_key` is the YAML key looked up in `config`. When present, the
-    helper emits `statement_fn(target, value_expr)` into the lambda body.
-    `target` is whatever the statement function needs to identify the
-    field (typically a setter name like `"set_position"` or a struct
-    member like `"position"`). `type_` is the C++ return type for
-    `cg.process_lambda` when the value is a user lambda.
-    """
-
-    conf_key: str
-    target: str
-    type_: object
-
-
-async def build_apply_lambda_action(
-    config: ConfigType,
-    action_id: ID,
-    template_arg: cg.TemplateArguments,
-    args: TemplateArgsType,
-    fields: tuple[ApplyField, ...],
-    prefix_args: list[tuple[object, str]],
-    statement_fn: Callable[[str, str], str],
-) -> MockObj:
-    """Fold configured fields into a single stateless apply lambda action.
-
-    Used by both `cover.control` and `cover.template.publish` (and shared
-    with the template/cover platform). Constants are emitted as flash
-    immediates; user lambdas are invoked inline so trigger args still flow.
-    Trigger arg types are normalized to `const std::remove_cvref_t<T> &`
-    to match the ApplyFn signature for any T (value, ref, or const-ref).
-    """
-    paren = await cg.get_variable(config[CONF_ID])
-    # Normalize trigger args to `const std::remove_cvref_t<T> &` so the
-    # apply lambda and any inner field lambdas (generated below via
-    # `process_lambda`) share one parameter spelling that's well-formed for
-    # any T.
-    normalized_args = [
-        (cg.RawExpression(f"const std::remove_cvref_t<{cg.safe_exp(t)}> &"), n)
-        for t, n in args
-    ]
-
-    fwd_args = ", ".join(name for _, name in args)
-    body_lines: list[str] = []
-    for field in fields:
-        if (value := config.get(field.conf_key)) is None:
-            continue
-        if isinstance(value, Lambda):
-            inner = await cg.process_lambda(
-                value, normalized_args, return_type=field.type_
-            )
-            value_expr = f"({inner})({fwd_args})"
-        else:
-            value_expr = str(cg.safe_exp(value))
-        body_lines.append(statement_fn(field.target, value_expr))
-
-    apply_args = [
-        *prefix_args,
-        *normalized_args,
-    ]
-    apply_lambda = LambdaExpression(
-        ["\n".join(body_lines)],
-        apply_args,
-        capture="",
-        return_type=cg.void,
-    )
-    return cg.new_Pvariable(action_id, template_arg, paren, apply_lambda)
-
-
 # CONF_STATE and CONF_POSITION are cv.Exclusive in the schema, so at most
 # one is present and both dispatch to set_position.
-_COVER_CONTROL_FIELDS: tuple[ApplyField, ...] = (
-    ApplyField(CONF_STOP, "set_stop", cg.bool_),
-    ApplyField(CONF_STATE, "set_position", cg.float_),
-    ApplyField(CONF_POSITION, "set_position", cg.float_),
-    ApplyField(CONF_TILT, "set_tilt", cg.float_),
+_COVER_CONTROL_FIELDS: tuple[actuator_component.ApplyField, ...] = (
+    actuator_component.ApplyField(CONF_STOP, "set_stop", cg.bool_),
+    actuator_component.ApplyField(CONF_STATE, "set_position", cg.float_),
+    actuator_component.ApplyField(CONF_POSITION, "set_position", cg.float_),
+    actuator_component.ApplyField(CONF_TILT, "set_tilt", cg.float_),
 )
 
 
@@ -393,7 +320,7 @@ async def cover_control_to_code(
     template_arg: cg.TemplateArguments,
     args: TemplateArgsType,
 ) -> MockObj:
-    return await build_apply_lambda_action(
+    return await actuator_component.build_apply_lambda_action(
         config=config,
         action_id=action_id,
         template_arg=template_arg,
